Question 2 naive multithreaded estimatePi
This code contains the commit for the naive implementation of
estimatePi.cpp. This code works by creating a thread for every single
point of data and then processing whether or not a point is within the
circle. This code is just used to illustrate that there is overhead
used when creating threads and that we need to be intelligent if we want
to achieve a proper speedup.