Maze Runner algorithm without visual backtracking
This commit contains the maze runner algorithm that is able to
recursively go through the maze. The only problem is that the code
doesn't display backtracking and just instantly returns to a different
position when it hits a dead end. Ideally the code would visually step
back through each step and then change paths. Either way it works.