Question 2 Commit
This commit completes all the code necessary for question 2. This
involves creating a shared memory which uses a single variable
processFlag to check whether or not a process is running. If the flag is
set the oindex, it's not running properly and the code will restart the
process. Otherwise code will continue.