Implements Customers waiting functionality
This code now makes it such that customers will wait for their meals to
be served before leaving the restaurant. This is accomplished by
creating a condition variable, and a servedDishCount variable that keep
track of the dishes serve. Currently, customers order an appetizer and a
main course, such that the condition variable waits until
servedDishCount == 2.