Poison Pill Implementation
This commit implements the poison pill implementation such that chefs
and servers will leave when all the customers are gone. This is
accomplished by creating a special poisonOrder which has item_id, and
customer_id = 666. Whenver a Chef or a Server encounters these items,
they will break from their respective loops and return.

In restaurant.cpp, when all the customers have left, we place poison
pills into the order queue based upon the number of chefs,and place
poison pills into the serve queue based upon the number of servers.
This is what causes the Chefs and Servers to receive poison pills in the
first place.